Fix subtract time (#5)

+++ src/Albatross.Expression/Functions/Date/SubtractTime.cs
@@ -36,7 +36,7 @@ namespace Albatross.Expression.Functions.Date
 			DateTime date1 = (DateTime)Convert.ChangeType(Operands[0].EvalValue(context), typeof(DateTime));
 			DateTime date2 = (DateTime)Convert.ChangeType(Operands[1].EvalValue(context), typeof(DateTime));
 
-			var result = date1.Subtract(date2).TotalMinutes;
+			var result = (long)date1.Subtract(date2).TotalMinutes;
 			return result;
 		}
 	}
